adding creating unit tests to orders funcionality and implementing the method to build a order from a cart

+++ src/Modules/SimplCommerce.Module.Shipments/Models/Requests/CalculatePackageFreightRequest.cs
@@ -3,7 +3,7 @@
     public class CalculatePackageFreightRequest
     {
         public string DestinationZipCode { get; set; }
-        public int CartId { get; set; }
+        public long CartId { get; set; }
         public string ShippingMethod { get; set; }        
     }
 }
